requirements.txt
# Corn Ji - Required Dependencies

# Core Framework
streamlit>=1.28.0

# Image Processing
Pillow>=10.0.0

# HTTP Requests (for Weather API)
requests>=2.31.0

# PDF Generation
reportlab>=4.0.0

# Optional: For better performance
numpy>=1.24.0

# Optional: For TensorFlow model (if using real AI model)
# tensorflow>=2.13.0
# tensorflow-macos>=2.13.0  # For Mac M1/M2
